Refactor dependency injection of SingleStreamEffectProvider from Set multibinding to a single Optional binding
